// apps/frontend/lib/api/generated/models/WorkExperience.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CredibilityHighlight } from "./CredibilityHighlight";
export type WorkExperience = {
	id?: string;
	companyName?: string;
	role?: string;
	startDate?: string;
	endDate?: string;
	description?: string;
	credibilityHighlights?: Array<CredibilityHighlight>;
};